🔀 Refactor NavComponent import and file structure
The commit refactors the import statement in the `profil/page.js` file to use the updated location of the `NavComponent`. It also deletes the old `Nav.component.js` file and adds a new `Nav.component.js` file with updated code. The changes ensure that the correct component is imported and used in the application.