✨ Added ModalBodyAllergens component
This commit adds the `ModalBodyAllergens` component, which is responsible for rendering a list of allergens in a modal body. It includes functionality to select and deselect allergens, as well as display their corresponding icons and names.

✨ Added ModalBodyIngredients component

This commit introduces the `ModalBodyIngredients` component, which displays a list of ingredients in a modal body. It allows users to select or deselect ingredients using checkboxes.

✨ Added ModalBodyMainContent component

This commit includes the `ModalBodyMainContent` component, which represents the main content section of a modal. It contains various input fields for dish name, description, image upload, ingredients selection, and price input.

✨ Added ModalHeaderContent component

This commit adds the `ModalHeaderContent` component that displays the header content of a modal. It dynamically renders different text based on whether allergen or ingredient updates are open or if it's editing a specific dish.

✨ Added ModalFooterBackAllergens component

This commit introduces the `ModalFooterBackAllergens` component that represents the footer section of a modal when updating allergens. It includes a button to go back to the previous step.

✨ Added ModalFooterBackIngredients component

This commit adds the `ModalFooterBackIngredients` component that represents the footer section of a modal when updating ingredients. It provides a button to go back to the previous step.

✨ Added ModalFooterMainContent component

This commit includes the `ModalFooterMainContent` component that represents the main footer section of a modal. It contains buttons for modifying allergens, selecting diets, canceling and closing, and saving changes and closing.