🔀 Add Get Categories API request
This commit adds a new file, `.bruno/Get Categories.bru`, which contains the code for making an HTTP GET request to retrieve categories from the API. The request includes headers for `Content-Type` and `Accept` with values of `application/json`. It also includes an authorization token in the `auth:bearer` section.